Support `-fno-exceptions`.
The only thing preventing the non-exception portions of GSL from
being used with `-fno-exceptions` is `gsl::narrow`. If we allow the
`-fno-exceptions` case to terminate with `gsl::Ensures` instead, the
library becomes usable in this case.

Note that AFAIK we can only automatically detect this case for Clang.
Non-Clang users will need to define `GSL_NO_EXCEPTIONS` themselves
when building.

Test: Made a test program using gsl::narrow, built and ran with and
      without -fno-exceptions. Can't add a test because unittest-cpp
      doesn't seem to support seem to support death tests, but `ctest`
      still passes for the default (`-fexceptions`) case.
Bug: http://b/24296825

Change-Id: Iae72dda03dbd82452fcc296897c73b913a54846a